requirements.txt
requests
python-dotenv
bind
plotly

Flask
Flask-SQLAlchemy
Flask-debugtoolbar
Flask-Migrate

